Refactor the link generating methods into its own module and created a class that is polymorphic to generate the links
